add permission to controllers

+++ app/Http/Controllers/CategoryController.php
@@ -11,8 +11,17 @@ class CategoryController extends Controller
     /**
      * Display a listing of the resource.
      */
+    public function checkPermission($permission)
+    {
+//        dd($permission, !auth()->user()->hasPermissionTo($permission));
+        if (!auth()->user()->hasPermissionTo($permission)) {
+            return to_route('admin.categories.index')->with('danger', 'You are not allowed to ' . $permission . '!');
+        }
+    }
+
     public function index()
     {
+
         $categories = Category::all();
         return view('admin.categories.index', compact('categories'));
     }
@@ -22,6 +31,11 @@ class CategoryController extends Controller
      */
     public function create()
     {
+//        $this->checkPermission('add category');
+        if (!auth()->user()->hasPermissionTo('add category')) {
+            return to_route('admin.categories.index')->with('danger', 'You are not allowed to add Category!');
+        }
+
         return view('admin.categories.create');
     }
 
@@ -30,6 +44,12 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth()->user()->hasPermissionTo('add category')) {
+            return to_route('admin.categories.index')->with('danger', 'You are not allowed to add Category!');
+        }
+
+        $this->checkPermission('add category');
+
         $image = substr($request->file('image')->store('public/categories'), 7);
         Category::create([
             'name' => $request->name,
@@ -52,6 +72,11 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('edit category')) {
+            return to_route('admin.categories.index')->with('danger', 'You are not allowed to edit Category!');
+        }
+
+//        $this->checkPermission('edit category');
         return view('admin.categories.edit', compact('category'));
     }
 
@@ -60,6 +85,11 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('edit category')) {
+            return to_route('admin.categories.index')->with('danger', 'You are not allowed to edit Category!');
+        }
+
+//        $this->checkPermission('edit category');
 
         $image = $category->image;
         if ($request->hasFile('image')) {
@@ -81,6 +111,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('delete category')) {
+            return to_route('admin.categories.index')->with('danger', 'You are not allowed to delete Category!');
+        }
+//        $this->checkPermission('delete category');
         $category->delete();
         return to_route('admin.categories.index')->with('warning', 'Category Deleted!');
